pyproject.toml
[tool.poetry]
name = "dribdat"
version = "0.5.0"
description = "Open source hackathon platform"
authors = ["Oleg Lavrovsky <oleg@datalets.ch>"]

[tool.poetry.dependencies]
python = ">=3.6,<4"
bleach = ">=3.1.4"
click = ">=6.7"
cssmin = ">=0.2.0"
email-validator = "*"
flask = ">=0.12.2"
flask-assets = ">=0.12"
flask-bcrypt = ">=0.7.1"
flask-caching = ">=1.3.3"
flask-cors = "*"
flask-hashing = ">=1.1"
flask-login = ">=0.4.0"
flask-migrate = ">=2.1.1"
flask-misaka = ">=0.4.1"
flask-sqlalchemy = ">=2.3.2"
flask-wtf = ">=0.14.2"
future = ">=0.16.0"
graphene = ">=1.4.1"
gunicorn = ">=19.7.1"
itsdangerous = ">=0.24"
jinja2 = ">=2.9.6"
jsmin = ">=2.2.2"
markupsafe = ">=1.0"
misaka = ">=2.1.0"
psycopg2-binary = ">=2.7.3.1"
pyquery = ">=1.2.17"
pytz = "*"
redis = ">=2.10.6"
requests = ">=2.18.4"
sqlalchemy = ">=1.1.14"
werkzeug = ">=0.12"
wtforms = ">=2.1"
flask-dance = { git = "https://github.com/loleg/flask-dance.git", branch = "master" }
email_validator = "^1.1.1"
python-dotenv = "^0.13.0"
flask-talisman = "^0.7.0"
urllib3 = "1.24.2"
pystache = "^0.5.4"

[tool.poetry.dev-dependencies]
factory-boy = "*"
flake8 = "*"
flake8-blind-except = "*"
flake8-debugger = "*"
flake8-docstrings = "*"
flake8-isort = "*"
flake8-quotes = "*"
Flask-DebugToolbar = "*"
isort = "*"
pep8-naming = "*"
pytest = "*"
WebTest = "*"
